Fix the parsing of the device responce
A reasonable logic has been implemented to parse the values in the
return string. Proper freeing of the values is implemented. Values of
voltage and current are handed back to the calling function.

There are debug-prints present to debug the behaviour of the funktion
later.

**Known Issues**

1. There seem to be instabilities with a sample size of 10. The reason for
this is unknown.

2. The serial read function is not working properly, because trailing and
leading non-print characters are also read from device