Change the communication with the power supply
Due to the slow communication with the power supply, we want to increase
the performance. After inspecting the problem, I found that the command
to obtain the values has an unnecessary overhead. We dropped about 60%
of the answer.

Now we use two dedicated commands to read the values for voltage and
current.

This increases the performance by a factor of 2.